+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this extension will be documented in this file.
 - Schema upload actions are now available in the Schemas view, as well as any Schema Registry shown
   in the Resources view. Schemas can also now be uploaded from an editor without needing to save to
   a file first.
+- Clicking a schema item in the Topics/Schemas view will now open the schema definition in a
+  read-only document without needing to click on the "View Schema" action.
 
 ## 0.22.1