+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this extension will be documented in this file.
 
 ## Unreleased
 
+## 0.22.1
+
+- Additional information is now provided when attempting to start/stop local resources and the
+  extension cannot communcate with the Docker engine API if the new `http.fetchAdditionalSupport`
+  setting is enabled.
+
+## 0.22.0
+
 ### Added
 
 - Ability to update basic topic configuration fields via webform